Bloodmoon and enderdragon in own files
Cleaned up code to remove junk from main.java and make it easier to find stuff.
MOST LIKELY WONT WORK FOR NOW
