+++ src/me/murrobby/igsq/expert/EnderDragon_Expert.java
@@ -1,0 +1,184 @@
+package me.murrobby.igsq.expert;
+
+import org.bukkit.Particle;
+import org.bukkit.Sound;
+import org.bukkit.World;
+import org.bukkit.World.Environment;
+import org.bukkit.entity.EnderDragon;
+import org.bukkit.entity.Entity;
+import org.bukkit.entity.EntityType;
+import org.bukkit.entity.Phantom;
+import org.bukkit.entity.Player;
+import org.bukkit.entity.EnderDragon.Phase;
+import org.bukkit.potion.PotionEffect;
+import org.bukkit.potion.PotionEffectType;
+
+import me.murrobby.igsq.Common;
+import me.murrobby.igsq.Main;
+
+import java.util.Random;
+
+public class EnderDragon_Expert
+{	
+	Main plugin;
+	Random random = new Random();
+	int enderDragonTask = -1;
+	final int taskID;
+	
+	public EnderDragon_Expert(Main plugin,int taskID) 
+	{
+		this.plugin = plugin;
+		this.taskID = taskID;
+		EnderDragonQuery();
+	}
+	private void EnderDragonQuery() 
+	{
+		enderDragonTask = plugin.scheduler.scheduleSyncRepeatingTask(plugin, new Runnable()
+    	{
+
+			@Override
+			public void run() 
+			{
+				EnderDragon();
+				if(Main_Expert.taskID != taskID || !Common.ExpertCheck()) 
+				{
+					plugin.scheduler.cancelTask(enderDragonTask);
+				}
+			} 		
+    	}, 0, 20);
+	}
+	private void EnderDragon() 
+	{
+		for(World selectedWorld : plugin.getServer().getWorlds()) 
+		{
+			try 
+			{
+				if(selectedWorld.getEnvironment() == Environment.THE_END) 
+				{
+					EnderDragon enderDragon = selectedWorld.getEnderDragonBattle().getEnderDragon();
+					if(enderDragon.getCustomName() == null) 
+					{
+						if(enderDragon.getPhase() == Phase.FLY_TO_PORTAL) 
+						{
+							for(Entity nearbyEntity : enderDragon.getNearbyEntities(50, 50, 50)) 
+							{
+								if(nearbyEntity.getType() == EntityType.PLAYER) 
+								{
+									if(random.nextInt(5)== 1) 
+									{
+										Player player = (Player) nearbyEntity;
+										Phantom phantom = (Phantom) selectedWorld.spawnEntity(enderDragon.getLocation(), EntityType.PHANTOM);
+										phantom.setCustomName("Expert Phantom Warrior");
+										phantom.setTarget(player);
+										phantom.setHealth(1);
+									}
+								}
+
+							}
+						}
+					}
+					else if(enderDragon.getCustomName().equalsIgnoreCase("True Expert Ender Dragon"))
+					{
+						if(enderDragon.getPhase() == Phase.LAND_ON_PORTAL || enderDragon.getPhase() == Phase.SEARCH_FOR_BREATH_ATTACK_TARGET) 
+						for(Entity nearbyEntity : enderDragon.getNearbyEntities(20, 20, 20)) 
+						{
+							if(nearbyEntity.getType() == EntityType.PLAYER) 
+							{
+								Player player = (Player) nearbyEntity;
+
+								if(enderDragon.getPhase() == Phase.LAND_ON_PORTAL) 
+								{
+									player.addPotionEffect(new PotionEffect(PotionEffectType.LEVITATION,79,4,true));
+									player.spawnParticle(Particle.DRAGON_BREATH,player.getLocation(), 20,.5,.5,.5,.25);
+								}
+								else if(enderDragon.getPhase() == Phase.SEARCH_FOR_BREATH_ATTACK_TARGET && player.hasPotionEffect(PotionEffectType.LEVITATION)) 
+								{
+									player.addPotionEffect(new PotionEffect(PotionEffectType.LEVITATION,10,128,true));
+									player.spawnParticle(Particle.END_ROD,player.getLocation(), 20,.5,.5,.5,.25);
+									player.playSound(player.getLocation(), Sound.ENTITY_DRAGON_FIREBALL_EXPLODE, 1,2);
+								}
+
+							}
+
+						}
+						if(enderDragon.getPhase() == Phase.FLY_TO_PORTAL) 
+						{
+							for(Entity nearbyEntity : enderDragon.getNearbyEntities(50, 50, 50)) 
+							{
+								if(nearbyEntity.getType() == EntityType.PLAYER) 
+								{
+									if(random.nextInt(2)== 1) 
+									{
+										Player player = (Player) nearbyEntity;
+										Phantom phantom = (Phantom) selectedWorld.spawnEntity(enderDragon.getLocation(), EntityType.PHANTOM);
+										phantom.setCustomName("Expert Phantom Warrior");
+										phantom.setTarget(player);
+										phantom.setHealth(3);
+									}
+								}
+
+							}
+						}
+						else if(enderDragon.getPhase() == Phase.DYING) 
+						{
+							for(Entity nearbyEntity : enderDragon.getNearbyEntities(100, 100, 100)) 
+							{
+								if(nearbyEntity instanceof Phantom) 
+								{
+									Phantom phantom = (Phantom) nearbyEntity;
+									if(phantom.getCustomName() != null && phantom.getCustomName().equalsIgnoreCase("Expert Phantom Warrior")) 
+									{
+										phantom.setHealth(0);
+									}
+								}
+							}
+						}
+						for(Entity nearbyEntity : enderDragon.getNearbyEntities(8, 8, 8)) 
+						{
+							if(nearbyEntity.getType() == EntityType.PLAYER) 
+							{
+
+								Player player = (Player) nearbyEntity;
+								player.addPotionEffect(new PotionEffect(PotionEffectType.BLINDNESS,59,0,false));
+
+							}
+
+						}
+						for(Entity nearbyEntity : enderDragon.getNearbyEntities(5, 5, 5)) 
+						{
+							if(nearbyEntity.getType() == EntityType.PLAYER) 
+							{
+
+								Player player = (Player) nearbyEntity;
+								player.addPotionEffect(new PotionEffect(PotionEffectType.WITHER,59,4,false));
+
+							}
+
+						}
+						if(enderDragon.getPhase() == Phase.STRAFING || enderDragon.getPhase() == Phase.CIRCLING) 
+						{
+							for(Entity nearbyEntity : enderDragon.getNearbyEntities(50, 50, 50)) 
+							{
+								if(nearbyEntity.getType() == EntityType.PLAYER) 
+								{
+
+									Player player = (Player) nearbyEntity;
+									player.addPotionEffect(new PotionEffect(PotionEffectType.BLINDNESS,59,0,false));
+									player.addPotionEffect(new PotionEffect(PotionEffectType.GLOWING,39,0,false));
+
+								}
+
+							}
+							enderDragon.addPotionEffect(new PotionEffect(PotionEffectType.GLOWING,39,0,false));
+						}
+					}
+
+				}
+			}
+			catch(Exception exception) 
+			{
+				System.out.println(exception.toString());
+			}
+		}
+	}
+}
