+++ src/me/murrobby/igsq/expert/EntityTargetEvent_Expert.java
@@ -37,7 +37,7 @@ public class EntityTargetEvent_Expert implements Listener
 	@EventHandler
 	public void EntityTarget_Expert(org.bukkit.event.entity.EntityTargetEvent event) 
 	{
-		if(!event.isCancelled())
+		if(Common.ExpertCheck() && !event.isCancelled())
 		{
 			if(event.getEntity() instanceof Phantom) 
 			{
