+++ src/me/murrobby/igsq/expert/PlayerBedEnterEvent_Expert.java
@@ -31,7 +31,7 @@ public class PlayerBedEnterEvent_Expert implements Listener
 	@EventHandler
 	public void PlayerBedEnter_Expert(org.bukkit.event.player.PlayerBedEnterEvent event) 
 	{
-		if(!event.isCancelled())
+		if(Common.ExpertCheck() && !event.isCancelled())
 		{
 			if(Common.getFieldBool(event.getPlayer().getWorld().getUID() + ".event.bloodmoon", "internal") && event.getPlayer().getWorld().getEnvironment() != Environment.NETHER && event.getPlayer().getWorld().getEnvironment() != Environment.THE_END)	
 			{
