+++ src/me/murrobby/igsq/main/BlockSpreadEvent_Main.java
@@ -1,28 +1,33 @@
-package me.murrobby.igsq.listeners;
+package me.murrobby.igsq.main;
 
 import org.bukkit.Bukkit;
+import org.bukkit.Server;
+import org.bukkit.block.BlockFace;
 import org.bukkit.entity.EntityType;
-import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
+import org.bukkit.util.Vector;
 
 import me.murrobby.igsq.Common;
 import me.murrobby.igsq.Main;
+import org.bukkit.Material;
+import java.util.Random;
 
 @SuppressWarnings("unused")
-public class EntityAirChangeEvent_Main implements Listener
+public class BlockSpreadEvent_Main implements Listener
 {
 	private Main plugin;
-	public EntityAirChangeEvent_Main(Main plugin)
+	Random random = new Random();
+	public BlockSpreadEvent_Main(Main plugin)
 	{
 		this.plugin = plugin;
 		Bukkit.getPluginManager().registerEvents(this, plugin);
 	}
 	
 	@EventHandler
-	public void EntityAirChange_Main(org.bukkit.event.entity.EntityAirChangeEvent event) 
+	public void BlockSpread_Main(org.bukkit.event.block.BlockSpreadEvent event) 
 	{
 		if(!event.isCancelled()) 
 		{
