+++ src/me/murrobby/igsq/main/CreatureSpawnEvent_Main.java
@@ -1,19 +1,18 @@
-package me.murrobby.igsq.listeners;
+package me.murrobby.igsq.main;
 
 import org.bukkit.Bukkit;
-import org.bukkit.Particle;
-import org.bukkit.World.Environment;
+import org.bukkit.attribute.Attribute;
+import org.bukkit.attribute.AttributeInstance;
+import org.bukkit.boss.BarColor;
 import org.bukkit.entity.Creeper;
-import org.bukkit.entity.Enderman;
+import org.bukkit.entity.EnderDragon;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.LivingEntity;
-import org.bukkit.entity.Phantom;
-import org.bukkit.entity.Player;
+import org.bukkit.entity.MagmaCube;
+import org.bukkit.entity.Slime;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import org.bukkit.event.entity.EntityDamageEvent.DamageCause;
-import org.bukkit.event.entity.EntityTargetEvent.TargetReason;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
 
@@ -24,22 +23,22 @@ import java.util.Random;
 
 
 @SuppressWarnings("unused")
-public class EntityTargetEvent_Main implements Listener
+public class CreatureSpawnEvent_Main implements Listener
 {
 	Random random = new Random();
 	private Main plugin;
-	public EntityTargetEvent_Main(Main plugin)
+	public CreatureSpawnEvent_Main(Main plugin)
 	{
 		this.plugin = plugin;
 		Bukkit.getPluginManager().registerEvents(this, plugin);
 	}
 	
 	@EventHandler
-	public void EntityTarget_Main(org.bukkit.event.entity.EntityTargetEvent event) 
+	public void CreatureSpawn_Main(org.bukkit.event.entity.CreatureSpawnEvent event) 
 	{
-		if(!event.isCancelled())
+		if(!event.isCancelled()) 
 		{
+			
 		}
 	}
-	
 }
