+++ src/me/murrobby/igsq/main/EntityAirChangeEvent_Main.java
@@ -1,11 +1,8 @@
-package me.murrobby.igsq.listeners;
+package me.murrobby.igsq.main;
 
 import org.bukkit.Bukkit;
-import org.bukkit.World.Environment;
-import org.bukkit.entity.Creeper;
-import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
-import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.potion.PotionEffect;
@@ -14,22 +11,18 @@ import org.bukkit.potion.PotionEffectType;
 import me.murrobby.igsq.Common;
 import me.murrobby.igsq.Main;
 
-import java.util.Random;
-
-
 @SuppressWarnings("unused")
-public class PlayerBedEnterEvent implements Listener
+public class EntityAirChangeEvent_Main implements Listener
 {
-	Random random = new Random();
 	private Main plugin;
-	public PlayerBedEnterEvent(Main plugin)
+	public EntityAirChangeEvent_Main(Main plugin)
 	{
 		this.plugin = plugin;
 		Bukkit.getPluginManager().registerEvents(this, plugin);
 	}
 	
 	@EventHandler
-	public void PlayerBedEnter_Main(org.bukkit.event.player.PlayerBedEnterEvent event) 
+	public void EntityAirChange_Main(org.bukkit.event.entity.EntityAirChangeEvent event) 
 	{
 		if(!event.isCancelled()) 
 		{
