+++ src/me/murrobby/igsq/main/EntityDamageByEntityEvent_Main.java
@@ -1,22 +1,22 @@
-package me.murrobby.igsq.listeners;
-
-import java.io.IOException;
-import java.time.LocalTime;
-import java.util.Collection;
+package me.murrobby.igsq.main;
 
 import org.bukkit.Bukkit;
+import org.bukkit.Particle;
 import org.bukkit.Sound;
 import org.bukkit.World.Environment;
-import org.bukkit.attribute.Attribute;
-import org.bukkit.boss.BarColor;
+import org.bukkit.entity.Arrow;
+import org.bukkit.entity.Creeper;
 import org.bukkit.entity.EnderDragon;
+import org.bukkit.entity.Enderman;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
+import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.MagmaCube;
+import org.bukkit.entity.Phantom;
 import org.bukkit.entity.Player;
+import org.bukkit.entity.Slime;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import org.bukkit.event.entity.EntityDamageEvent.DamageCause;
-import org.bukkit.event.entity.EntityPotionEffectEvent.Cause;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
 import org.bukkit.util.Vector;
@@ -24,23 +24,28 @@ import org.bukkit.util.Vector;
 import me.murrobby.igsq.Common;
 import me.murrobby.igsq.Main;
 
+import java.util.Random;
+
+
 @SuppressWarnings("unused")
-public class EntityDamageEvent_Main implements Listener
+public class EntityDamageByEntityEvent_Main implements Listener
 {
+	Random random = new Random();
 	private Main plugin;
-	public EntityDamageEvent_Main(Main plugin)
+	public EntityDamageByEntityEvent_Main(Main plugin)
 	{
 		this.plugin = plugin;
 		Bukkit.getPluginManager().registerEvents(this, plugin);
 	}
 	
 	@EventHandler
-	public void EntityDamage_Main(org.bukkit.event.entity.EntityDamageEvent event) 
+	public void EntityDamagedByEntity_Main(org.bukkit.event.entity.EntityDamageByEntityEvent event) 
 	{
 		if(!event.isCancelled()) 
 		{
 			
 		}
+			
 	}
 	
 }
