+++ src/me/murrobby/igsq/main/EntityTargetEvent_Main.java
@@ -1,13 +1,19 @@
-package me.murrobby.igsq.listeners;
+package me.murrobby.igsq.main;
 
 import org.bukkit.Bukkit;
+import org.bukkit.Particle;
 import org.bukkit.World.Environment;
 import org.bukkit.entity.Creeper;
+import org.bukkit.entity.Enderman;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Phantom;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
+import org.bukkit.event.entity.EntityDamageEvent.DamageCause;
+import org.bukkit.event.entity.EntityTargetEvent.TargetReason;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
 
@@ -18,22 +24,21 @@ import java.util.Random;
 
 
 @SuppressWarnings("unused")
-public class SlimeSplitEvent implements Listener
+public class EntityTargetEvent_Main implements Listener
 {
 	Random random = new Random();
 	private Main plugin;
-	public SlimeSplitEvent(Main plugin)
+	public EntityTargetEvent_Main(Main plugin)
 	{
 		this.plugin = plugin;
 		Bukkit.getPluginManager().registerEvents(this, plugin);
 	}
 	
 	@EventHandler
-	public void SlimeSplit_Main(org.bukkit.event.entity.SlimeSplitEvent event) 
+	public void EntityTarget_Main(org.bukkit.event.entity.EntityTargetEvent event) 
 	{
-		if(!event.isCancelled()) 
+		if(!event.isCancelled())
 		{
-			
 		}
 	}
