+++ src/me/murrobby/igsq/main/PlayerBedEnterEvent_Main.java
@@ -1,16 +1,11 @@
-package me.murrobby.igsq.listeners;
+package me.murrobby.igsq.main;
 
 import org.bukkit.Bukkit;
-import org.bukkit.attribute.Attribute;
-import org.bukkit.attribute.AttributeInstance;
-import org.bukkit.boss.BarColor;
+import org.bukkit.World.Environment;
 import org.bukkit.entity.Creeper;
-import org.bukkit.entity.EnderDragon;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.LivingEntity;
-import org.bukkit.entity.MagmaCube;
-import org.bukkit.entity.Slime;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.potion.PotionEffect;
@@ -23,22 +18,23 @@ import java.util.Random;
 
 
 @SuppressWarnings("unused")
-public class CreatureSpawnEvent_Main implements Listener
+public class PlayerBedEnterEvent_Main implements Listener
 {
 	Random random = new Random();
 	private Main plugin;
-	public CreatureSpawnEvent_Main(Main plugin)
+	public PlayerBedEnterEvent_Main(Main plugin)
 	{
 		this.plugin = plugin;
 		Bukkit.getPluginManager().registerEvents(this, plugin);
 	}
 	
 	@EventHandler
-	public void CreatureSpawn_Main(org.bukkit.event.entity.CreatureSpawnEvent event) 
+	public void PlayerBedEnter_Main(org.bukkit.event.player.PlayerBedEnterEvent event) 
 	{
 		if(!event.isCancelled()) 
 		{
 			
 		}
 	}
+	
 }
