+++ src/me/murrobby/igsq/main/PlayerJoinEvent_Main.java
@@ -1,4 +1,4 @@
-package me.murrobby.igsq.listeners;
+package me.murrobby.igsq.main;
 
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
