+++ src/me/murrobby/igsq/main/SlimeSplitEvent_Main.java
@@ -1,33 +1,35 @@
-package me.murrobby.igsq.listeners;
+package me.murrobby.igsq.main;
 
 import org.bukkit.Bukkit;
-import org.bukkit.Server;
-import org.bukkit.block.BlockFace;
+import org.bukkit.World.Environment;
+import org.bukkit.entity.Creeper;
+import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
+import org.bukkit.entity.LivingEntity;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
-import org.bukkit.util.Vector;
 
 import me.murrobby.igsq.Common;
 import me.murrobby.igsq.Main;
-import org.bukkit.Material;
+
 import java.util.Random;
 
+
 @SuppressWarnings("unused")
-public class BlockSpreadEvent implements Listener
+public class SlimeSplitEvent_Main implements Listener
 {
-	private Main plugin;
 	Random random = new Random();
-	public BlockSpreadEvent(Main plugin)
+	private Main plugin;
+	public SlimeSplitEvent_Main(Main plugin)
 	{
 		this.plugin = plugin;
 		Bukkit.getPluginManager().registerEvents(this, plugin);
 	}
 	
 	@EventHandler
-	public void BlockSpread_Main(org.bukkit.event.block.BlockSpreadEvent event) 
+	public void SlimeSplit_Main(org.bukkit.event.entity.SlimeSplitEvent event) 
 	{
 		if(!event.isCancelled()) 
 		{
